+++ src/Data/Puzzles/Compose.hs
@@ -65,6 +65,7 @@ handle f Labyrinth            = f R.labyrinth           D.labyrinth
 handle f Bahnhof              = f R.bahnhof             D.bahnhof
 handle f Cave                 = f R.cave                D.cave
 handle f AngleLoop            = f R.angleLoop           D.angleLoop
+handle f Shikaku              = f R.shikaku             D.shikaku
 
 -- | Handler that parses puzzle and an optional solution from a pair of
 --   corresponding YAML values, and renders both individually, optionally